remove engine, move prediction tests to test_tree
