+++ karoo_gp/__init__.py
@@ -2,7 +2,6 @@ from .pause import pause
 from .node_lib import NodeData, function_lib, get_nodes, get_function_node
 from .node import Node
 from .tree import Tree
-from .engine import Engine, NumpyEngine, TensorflowEngine
 from .population import Population
 from .base_class import BaseGP, RegressorGP, MultiClassifierGP, \
                         MatchingGP, ClassDecoder
